requirements.txt
uvicorn
python-dotenv
mysql-connector-python
pydantic
requests
pandas
fastapi
